+++ base/src/main/java/smile/math/matrix/Matrix.java
@@ -433,7 +433,7 @@ public class Matrix extends IMatrix {
 
     @Override
     public long size() {
-        return m * n;
+        return (long) m * n;
     }
 
     /**
